% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{abbreviation_to_state}
\alias{abbreviation_to_state}
\title{Convert two-letter symbol to state or territory name}
\usage{
abbreviation_to_state(x)
}
\arguments{
\item{x}{Vector of abbreviations}
}
\value{
Vector of state or territory names
}
\description{
Goes from "AZ" to "Arizona"
}
