% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{compile_grfp}
\alias{compile_grfp}
\title{Compile GRFP data from the external data in the package}
\usage{
compile_grfp()
}
\value{
A data.frame of all GRFP data
}
\description{
NSF releases GRFP information separately from other grant information, in two excel sheets per year: one for honorable mention, one for awards. These are downloaded as "xls" files but are actually "tsv" files. I have manually downloaded them for every year and stored them in `inst/extdata`. This function will compile them into one document. Package developers can then cache it in the package. For most users, just `data(grfp)` will be all you need to do to load in the data.
}
\examples{
#Not run
#grfp <- compile_grfp()
#usethis::use_data(grfp, overwrite=TRUE) # if you are updating the package as well
}
