% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{date_to_academic_semester}
\alias{date_to_academic_semester}
\title{Convert to academic semester}
\usage{
date_to_academic_semester(dates)
}
\arguments{
\item{date}{A date or vector of dates}
}
\value{
a character string or vector with the academic year (i.e. "2024 Fall")
}
\description{
Convert a date like 12/15/2024 to academic semester (2024 Fall). Note that to make it sort correctly, "Spring" is preceded by two spaces so that "2025  Spring" is alphabetically before "2025 Fall".
}
\examples{
dates <- as.Date(c("12/20/2024", "06/01/2021", "09/01/2021"), format="\%m/\%d/\%Y")
print(cbind(format(dates, "\%m/\%d/\%Y"), date_to_academic_semester(dates)))
}
