% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{date_to_academic_year}
\alias{date_to_academic_year}
\title{Convert to academic year}
\usage{
date_to_academic_year(dates)
}
\arguments{
\item{date}{A date or vector of dates}
}
\value{
a character string or vector with the academic year (i.e. "AY2024-2025")
}
\description{
Convert a date like 12/15/2024 to academic year (assumed to start on July 1)
}
\examples{
dates <- as.Date(c("12/20/2024", "06/01/2021", "09/01/2021"), format="\%m/\%d/\%Y")
print(cbind(format(dates, "\%m/\%d/\%Y"), date_to_academic_year(dates)))
}
