% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{flatten_list_columns}
\alias{flatten_list_columns}
\title{Collapse list columns}
\usage{
flatten_list_columns(grants)
}
\arguments{
\item{grants}{A data.frame of grant info}
}
\value{
A data.frame of grant info where each column is a "normal" vector: character, numeric, etc.
}
\description{
Some grants columns are lists as they may have multiple entries. This will collapse them with semicolons between each entry. Note that this will be SLOW on a large grants data.frame.
}
