% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\docType{data}
\name{grants}
\alias{grants}
\title{Grant information}
\format{
A data frame with one row per award and columns with award information
}
\usage{
grants
}
\description{
A dataset of NSF awards from its start until the package was last updated
}
\keyword{datasets}
