% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{load_nsf_grants}
\alias{load_nsf_grants}
\title{Load in NSF grants}
\usage{
load_nsf_grants(year_start = NA, year_end = NA, update_current_year = FALSE)
}
\arguments{
\item{year_start}{The earliest year of grants to incorporate (earliest available is 1960). Defaults to the current year}

\item{year_end}{The latest year of grants to incorporate. Defaults to the current year.}

\item{update_current_year}{If TRUE, and the year_end is the current year, it will automatically try to update the set of grants with the most recent ones from NSF's API. By default, it will just use the cached grants.}
}
\value{
A data.frame of the grants for the chosen year(s).
}
\description{
This will pull in grants from the rnsf releases on github. By default, it will only pull in the current calendar year's grants and then check to see if NSF has any new ones since the latest package release. You can change this behavior with arguments below.
}
\details{
To prevent the risk of including grants more than once, if update_current_year is true and the year_end is left NA or set to the current year, it will pull in all the grants from the current year, replacing the cached ones. This becomes slow by the end of the year.
}
\examples{
recent_grants_2018_2026 <- load_nsf_grants(year_start=2018, year_end=2026)
}
