% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_abstract_wordcloud}
\alias{nsf_abstract_wordcloud}
\title{Abstract wordcloud}
\usage{
nsf_abstract_wordcloud(grants = nsf_get_all(), ...)
}
\arguments{
\item{grants}{A data frame with grant information}
}
\description{
This follows the advice from http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know on making a word cloud
}
\details{
Make a word cloud of all the words in the abstracts
}
