% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_get_person}
\alias{nsf_get_person}
\title{Get all NSF grants for a person}
\usage{
nsf_get_person(first_name, middle_initial, last_name)
}
\arguments{
\item{first_name}{Just the first name}

\item{middle_initial}{Only one letter, no periods}

\item{last_name}{Only the last name}
}
\value{
A data frame with grant info
}
\description{
This will attempt to get all information for a person. Watch for multiple people with the same name
}
\examples{
# Let's check with my grants (mainly to show how to deal with weird characters, like apostrophes)
me <- nsf_get_person("Brian", "C", "O'Meara")
nsf_wordcloud(me$abstractText)
}
