% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_update_cached}
\alias{nsf_update_cached}
\title{Update cached information}
\usage{
nsf_update_cached()
}
\value{
A data.frame with the original grants data and appended new data
}
\description{
Adds grants after the last cached ones. However, it seemed to only add 60 grants in total; use nsf_update_cached_this_year() instead, or nsf_get_all() to get a new pull of every grant (takes days)
}
\examples{
# not running as it's slow
# grants <- nsf_update_cached()
# piggyback_one_year_grants(grants, format(Sys.Date(), "\%Y"))

}
