% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_update_this_year}
\alias{nsf_update_this_year}
\title{Update this year's information}
\usage{
nsf_update_this_year(grants)
}
\value{
A data.frame with the original grants data and appended new data
}
\description{
Adds grants, pulling all the ones from this year.
}
\examples{
# not running as it's slow
# grants <- nsf_update_this_year(grants)

}
