% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_wordcloud}
\alias{nsf_wordcloud}
\title{Text wordcloud}
\usage{
nsf_wordcloud(
  text = nsf_get_all()$abstractText,
  prune_words = c("will", "nfs"),
  max_words = 500,
  ...
)
}
\arguments{
\item{text}{A vector of text (for example, from grants$abstractText)}

\item{prune_words}{Other words you want to prune before plotting}

\item{max_words}{How many words at most to plot (it picks the most frequent ones)}

\item{...}{Arguments to the wordcloud function}
}
\description{
Create a wordcloud of text. This excludes common English words ("the", "and") but you can add your own to exclude as well. This uses the wordcloud package for plotting, and you can pass other arguments to that to make the plot prettier (see ?wordcloud::wordcloud)
This follows the advice from http://www.sthda.com/english/wiki/text-mining-and-word-cloud-fundamentals-in-r-5-simple-steps-you-should-know on making a word cloud
}
\details{
Make a word cloud of all the interesting words
}
\examples{
grants <- load_nsf_grants(year_start=2020, year_end=2024)
nsf_wordcloud(grants$abstractText[1:10])
}
