% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{piggyback_all_grants}
\alias{piggyback_all_grants}
\title{This will be used just once, to convert from the old style of storing as rdata to the new approach using piggyback}
\usage{
piggyback_all_grants(grants)
}
\description{
This will be used just once, to convert from the old style of storing as rdata to the new approach using piggyback
}
