% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{save_grant_csvs}
\alias{save_grant_csvs}
\title{Save CSVs of grants by decade}
\usage{
save_grant_csvs(grants, path)
}
\arguments{
\item{grants}{A data.frame of grant info}

\item{path}{Where to save the files}
}
\value{
Nothing, but the files will appear in the correct directory
}
\description{
This will save a CSV of all grants by decade (to make it easier to get the relevant files)
}
