% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\docType{data}
\name{grfp}
\alias{grfp}
\title{GRFP information}
\format{
A data frame with one row per award and columns with award information
}
\usage{
grfp
}
\description{
A dataset of NSF GRFP awards and honorable mentions from its start until the package was last updated
}
\keyword{datasets}
