% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_get_all}
\alias{nsf_get_all}
\title{Retrieve all NSF grant information, saving to a file}
\usage{
nsf_get_all(save_file = "NSFAllGrants.rda")
}
\arguments{
\item{save_file}{File to save results to while running}
}
\value{
A data frame with grant info
}
\description{
Retrieve all NSF grant information, saving to a file
}
