% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_return}
\alias{nsf_return}
\title{Retrieve table of grant information}
\usage{
nsf_return(keyword = NULL, zipcode = NULL, agency = NULL,
  verbose = TRUE, print_fields = print_fields_get(), save_file = NULL)
}
\arguments{
\item{keyword}{Optional keyword to search on}

\item{zipcode}{Optional zip code to limit search to (note that you should try both 5 and 9 digit). Note that there are often grants that lack zip code info}

\item{agency}{Agency to search for (NSF or NASA)}

\item{verbose}{If TRUE, outputs progress}

\item{print_fields}{Names of elements to return (see NSF API documentation)}

\item{save_file}{File to save results to while running}
}
\value{
A data frame with returned info
}
\description{
Retrieve table of grant information
}
\examples{
ants <- nsf_return(keyword="Formicidae")
nsf_wordcloud(ants$abstractText)

sanderson <- nsf_return(keyword="Sanderson")
mjs.PI <- sanderson[grepl("Michael J Sanderson",sanderson$pdPIName),] #note the lack of period for middle initial in the search string
mjs.CoPI <- sanderson[grepl("Michael J Sanderson",sanderson$coPDPI),]
mjs.grants <- rbind(mjs.PI, mjs.CoPI)
plot(x=lubridate::mdy(mjs.grants$startDate), y=mjs.grants$fundsObligatedAmt, pch=20, log="y", bty="n", xlab="Start date", ylab="Funding amount in US dollars")
}
