% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_update_cached}
\alias{nsf_update_cached}
\title{Update cached information}
\usage{
nsf_update_cached()
}
\value{
A data.frame with the original grants data and appended new data
}
\description{
Adds grants after the last cached ones. Once it is updated
}
\examples{
# not running as it's slow
# grants <- nsf_update_cached()
# usethis::use_data(grants, overwrite=TRUE) # if you are updating the package as well

}
