% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsf.R
\name{nsf_update_cached_this_year}
\alias{nsf_update_cached_this_year}
\title{Update cached information}
\usage{
nsf_update_cached_this_year()
}
\value{
A data.frame with the original grants data and appended new data
}
\description{
Adds grants, pulling all the ones from this year.
}
\examples{
# not running as it's slow
# grants <- nsf_update_cached_this_year()
# usethis::use_data(grants, overwrite=TRUE) # if you are updating the package as well

}
